Refactored and renamed. Depreciated Terminal
