+++ Terminal_Depreciated/puzzle_terminal.py
