+++ Terminal_Depreciated/terminal_main.py
